+++ R/control.R
@@ -31,7 +31,8 @@
 control_grid <- function(verbose = FALSE, allow_par = TRUE,
                          extract = NULL, save_pred = FALSE,
                          pkgs = NULL, save_workflow = FALSE,
-                         event_level = "first") {
+                         event_level = "first",
+                         parallel_over = "resamples") {
   # add options for  seeds per resample
 
   val_class_and_single(verbose, "logical", "control_grid()")
@@ -41,6 +42,7 @@ control_grid <- function(verbose = FALSE, allow_par = TRUE,
   val_class_and_single(event_level, "character", "control_grid()")
   val_class_or_null(pkgs, "character", "control_grid()")
   val_class_or_null(extract, "function", "control_grid()")
+  val_parallel_over(parallel_over, "control_grid()")
 
 
   res <- list(verbose = verbose,
@@ -49,7 +51,8 @@ control_grid <- function(verbose = FALSE, allow_par = TRUE,
               save_pred = save_pred,
               pkgs = pkgs,
               save_workflow = save_workflow,
-              event_level = event_level)
+              event_level = event_level,
+              parallel_over = parallel_over)
 
   class(res) <- c("control_grid", "control_resamples")
   res
@@ -105,6 +108,19 @@ control_resamples <- control_grid
 #'   This argument is passed on to yardstick metric functions when any type
 #'   of class prediction is made, and specifies which level of the outcome
 #'   is considered the "event".
+#' @param parallel_over A single string containing either `"resamples"` or
+#'   `"everything"` describing how to use parallel processing.
+#'
+#'   If `"resamples"`, then tuning will be performed in parallel over resamples
+#'   alone. Within each resample, the preprocessor (i.e. recipe or formula) is
+#'   processed once, and is then reused across all models that need to be fit.
+#'
+#'   If `"everything"`, then tuning will be performed in parallel at two levels.
+#'   An outer parallel loop will iterate over resamples. Additionally, an
+#'   inner parallel loop will iterate over all unique combinations of
+#'   preprocessor and model tuning parameters for that specific resample. This
+#'   will result in the preprocessor being re-processed multiple times, but
+#'   can be faster if that processing is extremely fast.
 #'
 #' @details
 #'
@@ -138,7 +154,8 @@ control_bayes <-
            pkgs = NULL,
            save_workflow = FALSE,
            save_gp_scoring = FALSE,
-           event_level = "first") {
+           event_level = "first",
+           parallel_over = "resamples") {
     # add options for seeds per resample
 
     val_class_and_single(verbose, "logical", "control_bayes()")
@@ -152,6 +169,7 @@ control_bayes <-
     val_class_and_single(time_limit, c("logical", "numeric"), "control_bayes()")
     val_class_or_null(pkgs, "character", "control_bayes()")
     val_class_and_single(event_level, "character", "control_bayes()")
+    val_parallel_over(parallel_over, "control_bayes()")
 
     if (!is.infinite(uncertain) && uncertain > no_improve) {
       cli::cli_alert_warning(
@@ -171,7 +189,8 @@ control_bayes <-
         pkgs = pkgs,
         save_workflow = save_workflow,
         save_gp_scoring = save_gp_scoring,
-        event_level = event_level
+        event_level = event_level,
+        parallel_over = parallel_over
       )
 
     class(res) <- "control_bayes"
@@ -181,4 +200,12 @@ control_bayes <-
 print.control_bayes <- function(x, ...) {
   cat("bayes control object\n")
   invisible(x)
+}
+
+# ------------------------------------------------------------------------------
+
+val_parallel_over <- function(parallel_over, where) {
+  val_class_and_single(parallel_over, "character", where)
+  rlang::arg_match0(parallel_over, c("resamples", "everything"), "parallel_over")
+  invisible(NULL)
 }